backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
pydantic==2.7.0
pymongo==4.6.1
beautifulsoup4==4.12.3
requests==2.32.0
anyio==4.4.0
google-generativeai
python-dotenv
